refactor spv
